+++ src/include/oeis_list.hpp
@@ -20,6 +20,5 @@ class OeisList {
   static void mergeMap(const std::string& file_name,
                        std::map<size_t, int64_t>& map);
 
- private:
-  static std::string LISTS_HOME;
+  static const std::string INVALID_MATCHES_FILE;
 };